IOTDEV-88: Trying to get organized in the assets folder.
