+++ assets/_custom-variables.scss
@@ -44,56 +44,7 @@ $h5-font-size: $font-size-base * 1.1;
 /* Set sizeing variable */
 $sizer: 1rem;
 
-/* Import all of Bootstrap */
-@import '~bootstrap/scss/bootstrap';
+@import "~bootstrap/scss/_functions.scss";
+@import "~bootstrap/scss/_mixins.scss";
+@import '~bootstrap/scss/_variables.scss';
 
-/* Import partials */
-@import 'onboarding';
-@import 'navigation';
-@import 'explore';
-@import 'text';
-@import 'carousel';
-
-/* Add generic styling outside Bootstrap */
-body {
-  background-blend-mode: luminosity;
-  background-position: center;
-  background-size: cover;
-  min-height: 100vh;
-  padding: $sizer / 2;
-}
-
-body > .container-fluid {
-  box-shadow: $dropdown-box-shadow;
-  min-height: calc(100vh - #{$sizer});
-}
-
-table {
-  .number {
-    text-align: right;
-  }
-}
-
-#chart {
-    width: 100%;
-    height: 500px;
-}
-
-
-// @TODO: Design
-.badge-log-entry-type {
-  @include badge-variant($primary-300);
-}
-
-.badge-log-entry-type-user {
-  @include badge-variant($primary-400);
-}
-
-.badge-log-entry-type-system {
-  @include badge-variant($primary-200);
-}
-
-.log-view-content {
-  max-height: 500px;
-  overflow-y: scroll;
-}
