--- assets/css/_explore.scss
@@ -1,10 +1,0 @@
-.mission-dot {
-  background-color: $primary;
-  border-radius: 50%;
-  box-shadow: $box-shadow-sm;
-  display: inline-block;
-  height: $spacer;
-  margin: 0;
-  padding: 0;
-  width: $spacer;
-}
