--- assets/css/_navigation.scss
@@ -1,18 +1,0 @@
-.page-navigation {
-  bottom: $sizer;
-  left: 0;
-  position: fixed;
-  right: 0;
-}
-
-.close {
-  position: fixed;
-  right: $sizer;
-  top: $sizer;
-}
-
-.btn-add-new-mission {
-  bottom: $spacer;
-  position: fixed;
-  right: $spacer;
-}
