--- assets/css/_text.scss
@@ -1,8 +1,0 @@
-
-.text-primary-faded {
-  color: $primary-faded;
-}
-
-.badge {
-  font-weight: 400;
-}
