+++ assets/default/default.scss
@@ -1,0 +1,7 @@
+@import "../_custom-variables.scss";
+
+.btn-add-new-mission {
+  bottom: $spacer;
+  position: fixed;
+  right: $spacer;
+}
