+++ assets/default/index.js
@@ -1,0 +1,14 @@
+/* Import base styling */
+import '../base.scss'
+
+/* Import this components styling */
+import './default.scss'
+
+/* Import FontAwesome icons */
+import '@fortawesome/fontawesome-free/js/all'
+
+/* Require roboto typeface */
+require('typeface-roboto')
+require('jquery')
+require('popper.js')
+require('bootstrap')
