+++ assets/mission/LogEntry.jsx
