+++ assets/mission/LogView.jsx
