--- assets/mission/css/app.scss
@@ -1,39 +1,0 @@
-@import "~bootstrap/scss/mixins/_badge.scss";
-
-// // html, body, #app, #app > div {
-// //     padding: 0;
-// //     height: 100%;
-// //     overflow: hidden;
-// // }
-
-// #chart {
-//     height: 500px;
-// }
-
-// // .log-entry-alert {
-// //     @extend .alert-danger;
-// // }
-
-// // .btn-add-log-entry,
-// // .btn-add-annotation {
-// //     @extend .btn-secondary;
-// // }
-
-// // .badge-log-entry-type-user {
-// //   @include badge-variant($primary-400);
-// // }
-
-// // % if type == 'User' %}primary-400{% elseif type == 'system'%}primary-200{% else %}primary-300{% endif %}
-
-
-// .badge-log-entry-type {
-//   @include badge-variant($primary-300);
-// }
-
-// .badge-log-entry-type-user {
-//   @include badge-variant($primary-400);
-// }
-
-// .badge-log-entry-type-system {
-//   @include badge-variant($primary-200);
-// }
