--- assets/mission/css/index.scss
@@ -1,14 +1,0 @@
-@import '../../../node_modules/leaflet/dist/leaflet';
-
-#map {
-    height: 400px;
-    width: 100%;
-}
-
-.map-marker-icon {
-  width: 12px;
-  height: 12px;
-  background: rgba(0, 127, 0, .75);
-  border-radius: 6px;
-  // @TODO: Design: Add a fontawesome marker or something …
-}
