+++ assets/mission_theme/mission_theme.scss
@@ -1,3 +1,6 @@
+@import "../custom-variables";
+@import "~bootstrap/scss/carousel";
+
 .carousel-indicators {
   position: static;
