+++ assets/onboarding/index.js
@@ -1,0 +1,14 @@
+/* Import base styling */
+import '../base.scss'
+
+/* Import this components styling */
+import './onboarding.scss'
+
+/* Import FontAwesome icons */
+import '@fortawesome/fontawesome-free/js/all'
+
+/* Require roboto typeface */
+require('typeface-roboto')
+require('jquery')
+require('popper.js')
+require('bootstrap')
