+++ assets/onboarding/map_background@2x.png
