+++ assets/onboarding/onboarding.scss
@@ -1,3 +1,5 @@
+@import "../custom-variables";
+
 .onboarding {
   h1 {
     margin-left: $sizer * 2;
