+++ issues.md
@@ -1,0 +1,12 @@
+# Issues
+
+## CMS admin
+
+* When you have added a image to a page, saves, and revisit the page, the file name is no longer present in the input field. This means that there is no longer any clue left to the user, to show that a image is already present.
+
+## Todo
+
+* Add log message
+* Add new sensor
+* Edit sensors
+*
