--- templates/app/index.html.twig
@@ -1,7 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}IoT Crawler App{% endblock %}
-
-{% block content %}
-    App!
-{% endblock %}
