--- templates/app/mission/new.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body_class %}new-mission{% endblock %}
-
-{% block background_image %}style="background-image: url({{ asset('build/images/map_background.png') }});"{% endblock %}
-
-{% block content %}
-    <div class="container-fluid bg-white">
-        <article class="mission py-3">
-            {#% include "mock-up/_partials/_nav_close.html.twig" with {'nav_path': '/mock-up/explore' }%#}
-
-            <h1>{{ "Create mission"|trans }}</h1>
-            <p class="text-primary-faded text-uppercase small">{{ theme.title }}
-                {#<a href="{{ path('app_theme_show', {theme: theme.id}) }}" class="ml-3">Show theme</a>#}
-            </p>
-
-            <form>
-                <div class="form-group">
-                    <label for="missionTitle">{{ "Mission title"|trans }}</label>
-                    <input type="text" class="form-control" id="missionTitle" aria-describedby="missionTitleHelp">
-                    <small id="missionTitleHelp" class="form-text text-muted">{{ "Give your mission a easy recognisable titel"|trans }}</small>
-                </div>
-                <div class="form-group">
-                    <label for="missionDescription">{{ "Mission description"|trans }}</label>
-                    <textarea class="form-control" id="missionDescription" aria-describedby="missionDescriptionHelp" rows="3"></textarea>
-                    <small id="missionDescriptionHelp" class="form-text text-muted">{{ "Give a short description of what you want to achive."|trans }}</small>
-                </div>
-                <div class="form-group">
-                    <label for="missionLocation">{{ "Mission location"|trans }}</label>
-                    <input type="text" class="form-control" id="missionLocation" aria-describedby="missionLocationHelp">
-                    <small id="missionLocationHelp" class="form-text text-muted">{{ "Write the address or primary location of yor mission"|trans }}</small>
-                </div>
-
-                <button type="submit" class="btn btn-success btn-block">Submit</button>
-            </form>
-
-        </article>
-    </div>
-{% endblock %}
