--- templates/app/themes.html.twig
@@ -1,46 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title 'Mission theme'|trans %}
-
-{% block body_class 'new-mission' %}
-
-{% block background_image %}style="background-image: url({{ asset('build/images/map_background.png') }});"{% endblock %}
-
-{% block content %}
-    <div class="container-fluid bg-white">
-        <article class="mission py-3">
-            {#% include "_partials/_nav_close.html.twig" with {'nav_path': '/mock-up/explore' }%#}
-
-            <h1>{{ 'Mission theme'|trans }}</h1>
-            <p>{{ 'Before you can create a datamission you need to select a mission theme'|trans }}</p>
-
-            <div id="carouselMissionThemes" class="carousel slide" data-ride="false" data-interval="false">
-
-                <div class="carousel-inner bg-success-300 p-3 text-center">
-                    {% for theme in themes %}
-
-                        <div class="carousel-item {{ loop.first ? 'active' }} my-3">
-                            <h1>{{ theme.title }}</h1>
-
-                            <span class="fa-8x">
-                                <span class="fa-layers fa-fw">
-                                    <i class="fas fa-circle text-white"></i>
-                                    <i class="fas fa-invers fa-{{ theme.icon }}" data-fa-transform="shrink-6"></i>
-                                </span>
-                            </span>
-
-                            <p class="description">{{ theme.description }}</p>
-                            <a href="{{ path('mission_new', {theme: theme.id}) }}" class="btn btn-success">{{ "Create new mission"|trans }}</a>
-                        </div>
-                    {% endfor %}
-                </div>
-                <ol class="carousel-indicators my-3">
-                    {% for theme in themes %}
-                        <li data-target="#carouselMissionThemes" data-slide-to="{{ loop.index0 }}" class="{{ loop.first ? 'active' }}"></li>
-                    {% endfor %}
-                </ol>
-            </div>
-
-        </article>
-    </div>
-{% endblock %}
