+++ templates/base.html.twig
@@ -10,7 +10,7 @@
 			{% endblock %}
 		</title>
 		{% block stylesheets %}
-			{{ encore_entry_link_tags('app') }}
+			{{ encore_entry_link_tags('base') }}
 		{% endblock %}
 	</head>
 	<body class="{% block body_class %}{% endblock %}" {% block background_image %}{% endblock %}>
@@ -18,7 +18,7 @@
             {% block content %}{% endblock %}
 		{% endblock %}
 		{% block javascripts %}
-			{{ encore_entry_script_tags('app') }}
+			{{ encore_entry_script_tags('base') }}
 		{% endblock %}
 	</body>
 </html>
