+++ templates/mission/new.html.twig
@@ -20,10 +20,10 @@
 
 {% block stylesheets %}
     {{ parent() }}
-    {{ encore_entry_link_tags('mission/edit') }}
+    {{ encore_entry_link_tags('mission') }}
 {% endblock %}
 
 {% block javascripts %}
     {{ parent() }}
-	{{ encore_entry_script_tags('mission/edit') }}
+	{{ encore_entry_script_tags('mission') }}
 {% endblock %}
