+++ templates/mission_theme/show.html.twig
@@ -2,6 +2,11 @@
 
 {% block title %}MissionTheme{% endblock %}
 
+{% block stylesheets %}
+    {{ parent() }}
+    {{ encore_entry_link_tags('mission-theme') }}
+{% endblock %}
+
 {% block body %}
     <h1>MissionTheme</h1>
 
@@ -47,4 +52,9 @@
     <a href="{{ path('mission_theme_edit', {'id': mission_theme.id}) }}">edit</a>
 
     {{ include('mission_theme/_delete_form.html.twig') }}
+{% endblock %}
+
+{% block javascripts %}
+    {{ parent() }}
+    {{ encore_entry_script_tags('mission-theme') }}
 {% endblock %}
