+++ webpack.config.js
@@ -23,10 +23,10 @@ Encore
  * Each entry will result in one JavaScript file (e.g. app.js)
  * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
  */
-  .addEntry('app', './assets/js/app.js')
-  .addEntry('mission/app', './assets/mission/js/app.jsx')
-  .addEntry('mission/index', './assets/mission/js/index.js')
-  .addEntry('mission/edit', './assets/mission/js/edit.js')
+  .addEntry('base', './assets/base.js')
+  .addEntry('mission', './assets/mission/index.js')
+  .addEntry('mission-theme', './assets/mission_theme/index.js')
+  .addEntry('onboarding', './assets/onboarding/index.js')
 
 // When enabled, Webpack "splits" your files into smaller pieces for greater optimization.
   .splitEntryChunks()
@@ -76,12 +76,12 @@ Encore
 // .addEntry('admin', './assets/js/admin.js')
 
 // Enable copy of static assets
-  .copyFiles({
-    from: './assets/images',
-    // if versioning is enabled, add the file hash too
-    to: 'images/[path][name].[hash:8].[ext]',
-    // only copy files matching this pattern
-    pattern: /\.(png|jpg|jpeg)$/
-  })
+  // .copyFiles({
+  //   from: './assets/images',
+  //   // if versioning is enabled, add the file hash too
+  //   to: 'images/[path][name].[hash:8].[ext]',
+  //   // only copy files matching this pattern
+  //   pattern: /\.(png|jpg|jpeg)$/
+  // })
 
 module.exports = Encore.getWebpackConfig()
